requirements.txt
agate==1.7.1
attrs==23.1.0
Babel==2.13.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
dbt-core==1.6.6
dbt-extractor==0.4.1
dbt-postgres==1.6.6
dbt-semantic-interfaces==0.2.2
hologram==0.0.16
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
leather==0.3.4
Logbook==1.5.3
MarkupSafe==2.1.3
mashumaro==3.8.1
minimal-snowplow-tracker==0.0.2
more-itertools==8.14.0
msgpack==1.0.7
networkx==3.2
packaging==23.2
parsedatetime==2.6
pathspec==0.11.2
protobuf==4.24.4
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==1.10.13
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.8.0
urllib3==1.26.18
zipp==3.17.0
